Update script.js
Several big changes.

Added saving and reading of JSON file. So it saves the "application" and a user can upload it. Including the geojson data and user data.
Also fixed minor bugs and added more comments in code. 
Also changed size of infobar on the right.
Also added resize capabilities for when window is resized